+++ CHANGELOG.md
@@ -53,7 +53,7 @@
 - New #67: Add `editGeneralForumTopic`, `closeGeneralForumTopic`, `reopenGeneralForumTopic`, `hideGeneralForumTopic`
   and `unhideGeneralForumTopic` methods.
 - New #68: Add `answerCallbackQuery` method.
-- New #68: Add `UserChatBoosts` type.
+- New #69: Add `getUserChatBoosts` method and `UserChatBoosts` type.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.